requirements.txt
fastapi
pillow
python-dotenv
